tui: fix clang-tidy's findings in tui/
Running CI's clang-tidy 18 configuration over tui/ turned up about 140
findings besides include-what-you-use. They are fixed:
- Ignored results of fprintf, fclose, fseek, and the like are cast to
  (void), as in src/.
- fopen and open now use close-on-exec ("e" / O_CLOEXEC).
- Nested ternaries become if/else, or switch helpers for the
  play-setup labels.
- Ternaries mixing char and int get one explicit char cast.
- Redundant casts, stores, and branches are removed; multi-variable
  declarations are split; widening multiplications and literal
  suffixes are fixed.
- A #define constant becomes an enum.
- atoi and atof become strtol and strtod.
- RackTileCache fields are reordered to drop padding.

The static analyzer found three NULL paths:
- tui_game_state_init allows a NULL error buffer, but copy_error
  wrote to it anyway.
- render_modal_ex checked items for NULL in one place and indexed it
  unchecked in another.
- The tile composers indexed a NULL letter distribution.
Each now checks up front.

+++ tui/analysis_rows.c
@@ -399,7 +399,7 @@ void populate_frame_analysis_rows(TuiGameState *state) {
     struct timespec now;
     clock_gettime(CLOCK_MONOTONIC, &now);
     const long since_ms = (long)(now.tv_sec - last_build.tv_sec) * 1000L +
-                          (long)(now.tv_nsec - last_build.tv_nsec) / 1000000L;
+                          (now.tv_nsec - last_build.tv_nsec) / 1000000L;
     if (last_build_count > 0 && since_ms >= 0 && since_ms < 100) {
       state->last_rendered_analysis_row_count = last_build_count;
       return;
